+++ CHANGES.md
@@ -6,6 +6,7 @@
 ### Changes ###
 
 * MTBuild v0.1.0 is a big, API-breaking change.
+* MTBuild now requires Ruby >= 2.0.0.
 * MTBuild now overrides the rake application name to display "mtbuild" as the
   application name.
 * MTBuild now offers a '--super-dry-run' command line option to perform a dry
@@ -15,11 +16,11 @@
 * Default tasks can no longer be added to projects. You can only add default
   tasks to workspaces. If you need default tasks for a project, simply include
   a lightweight workspace that exists solely to specify default tasks.
-* The "clobber" task is now gone. MTbuild workspaces now generate a top-level 
+* The "clobber" task is now gone. MTbuild workspaces now generate a top-level
   "clean" task that cleans all projects. Additionally, each project provides
   its own "clean" task for cleaning just one project at a time. The "clean"
   tasks remove all intermediate files and final output, so they behave like
-  the older Rake "clobber" task. 
+  the older Rake "clobber" task.
 * Added the "gpp" toolchains. These are similar to the gcc toolchains, but they
   invoke g++ instead of gcc.